pyproject.toml
[tool.poetry]
name = "seedweb"
version = "0.1.0"
description = ""
authors = ["Tom Camp <tomcamp@protonmail.ch>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
flask-wtf = "^1.2.1"
flask-restful = "^0.3.10"
flask-cors = "^4.0.0"
Jinja2 = "^3.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
